backend/requirements.txt
# backend/requirements.txt

# Core FastAPI framework
fastapi==0.111.0
uvicorn[standard]==0.29.0 # Includes websockets support

# Face Recognition & Image Processing
face_recognition==1.3.0
dlib==20.0.0 # Dependency for face_recognition
numpy==1.26.4
opencv-python-headless==4.9.0.80 # Use headless if no GUI needed

# Database & Encryption
pymongo==4.7.2
cryptography==42.0.7
python-dotenv==1.0.1 # For loading .env files

# Real-time Communication
python-socketio==5.11.2
# Note: uvicorn[standard] already includes 'websockets'

# Deepfake & AI (Dependencies for insightface, torch, etc.)
insightface==0.7.3 # Install insightface if using RetinaFace
onnxruntime==1.17.3 # Dependency for insightface CPU
torch==2.3.0        # If using PyTorch models (like MobileNetV3)
torchvision==0.18.0 # If using PyTorch models

# Notifications (Optional, if implementing SMS)
twilio==8.11.1 # Example SMS provider

# Other Utilities
pydantic==2.7.1 # For data validation models
python-multipart==0.0.9 # For handling file uploads in FastAPI
